+++ trunk/src/database/eventinfo.cpp
@@ -120,7 +120,7 @@ void EventInfo::update()
     // Clean previous statistics
     reset();
 
-    for(int i = 0; i < m_database->count(); ++i)
+    for(int i = 0; i < (int) m_database->count(); ++i)
     {
         if(index->valueIndexFromTag(TagNameEvent, i) != event)
         {
